Fix: ETL convert column from string to numeric  issues

+++ data/process_data.py
@@ -19,14 +19,12 @@ def clean_data(df):
 
     #Convert category values to just numbers 0 or 1
     for column in categories:
-    # set each value to be the last character of the string
+        # set each value to be the last character of the string
         categories[column] = categories[column].apply(lambda x:x[-1:])
-
+        # convert column from string to numeric
+        categories[column] = categories[column].astype(int)
     #replace value 2 by 1 as anything above 1 is true
     categories=categories.replace(2,1)
-
-    # convert column from string to numeric
-    categories[column] = categories[column].astype(int)
 
     #Replace categories column in df with new category columns
     # drop the original categories column from `df`
